utils/src/main/java/com/example/utils/Result.kt
package com.example.utils

import kotlinx.coroutines.flow.Flow
import kotlinx.coroutines.flow.catch
import kotlinx.coroutines.flow.map
import kotlinx.coroutines.flow.onStart

sealed interface Result<out T> {
    data class Success<T>(val data: T) : Result<T>
    object Loading : Result<Nothing>
    data class Error(val error: Throwable) : Result<Nothing>
}

@Suppress("UNCHECKED_CAST")
fun <T> Flow<T>.asResult(): Flow<Result<T>> =
    this
        .map<T, Result<T>> { Result.Success(it) }
        .onStart { emit(Result.Loading) }
        .catch { t -> emit(Result.Error(t)) }
